move crate-level types and constants
