+++ bfd-async/Cargo.toml
@@ -8,6 +8,7 @@ anyhow.workspace = true
 bfd.workspace = true
 mg-common.workspace = true
 mg-api-types.workspace = true
+rand.workspace = true
 rdb.workspace = true
 slog.workspace = true
 slog-error-chain.workspace = true
