+++ bfd-async/src/dispatcher/end_to_end.rs
@@ -18,9 +18,9 @@ use super::ListenerTask;
 use super::SharedSessions;
 use crate::AddPeerError;
 use crate::AddPeerRequest;
+use crate::DEFAULT_DETECT_MULTIPLIER;
 use crate::Daemon;
 use crate::wait_for_condition;
-use bfd::DEFAULT_DETECT_MULTIPLIER;
 use mg_api_types::bfd::BfdPeerState;
 use mg_api_types::bfd::SessionMode;
 use mg_common::lock;
