+++ bfd-async/src/lib.rs
@@ -8,6 +8,8 @@ use std::io;
 use std::net::IpAddr;
 use std::net::SocketAddr;
 use std::num::NonZeroU8;
+use std::sync::atomic::AtomicU64;
+use std::time::Duration;
 
 mod daemon;
 mod dispatcher;
@@ -22,11 +24,17 @@ pub use daemon::Daemon;
 pub use dispatcher::ListenerShutdownHandle;
 pub use session::Session;
 
+const COMPONENT_BFD: &str = "bfd";
+const MOD_DAEMON: &str = "daemon";
+const UNIT_PEER: &str = "peer";
+
+const DEFAULT_BFD_TTL: u32 = 255;
+
 /// Default detection threshold multiplier.
 ///
 /// Three seems to be a common choice for other implementations. Without
 /// intuition for or against this default, follow suit.
-pub const DEFAULT_DETECT_MULTIPLIER: NonZeroU8 = NonZeroU8::new(3).unwrap();
+const DEFAULT_DETECT_MULTIPLIER: NonZeroU8 = NonZeroU8::new(3).unwrap();
 
 /// Errors from attempting to add a new BFD peer.
 #[derive(Debug, thiserror::Error)]
@@ -86,6 +94,85 @@ impl From<BfdPeerConfig> for AddPeerRequest {
             required_rx_micros,
             detection_threshold,
             mode,
+        }
+    }
+}
+
+#[derive(Default)]
+pub struct SessionCounters {
+    pub control_packets_sent: AtomicU64,
+    pub control_packet_send_failures: AtomicU64,
+    pub control_packets_received: AtomicU64,
+    pub admin_down_status_received: AtomicU64,
+    pub down_status_received: AtomicU64,
+    pub init_status_received: AtomicU64,
+    pub up_status_received: AtomicU64,
+    pub unknown_status_received: AtomicU64,
+    pub transition_to_init: AtomicU64,
+    pub transition_to_down: AtomicU64,
+    pub transition_to_up: AtomicU64,
+    pub timeout_expired: AtomicU64,
+    pub message_receive_error: AtomicU64,
+    pub unexpected_message: AtomicU64,
+}
+
+/// Information about a BFD peer.
+#[derive(Debug, Clone, Copy)]
+pub struct PeerInfo {
+    /// The interval at which the peer would _like_ to receive BFD control
+    /// packets.
+    pub desired_min_tx: Duration,
+
+    /// This is the minimum interval between received BFD Control packets that
+    /// this system is capable of supporting.
+    pub required_min_rx: Duration,
+
+    /// A unique identifer for the peer. This structure is used to keep track
+    /// of remote peer information as well as our own. The remote peer
+    /// generates their own discriminator. When a peer state machine is first
+    /// started, we generate our discriminator with
+    /// `PeerInfo::with_random_discriminator`
+    pub discriminator: u32,
+
+    /// Whether or not the peer is requesting demand mode. This means
+    /// unsolicited BFD control packets will not be sent. The only control
+    /// packets sent will be in response to control packets received with the
+    /// poll flag set.
+    pub demand_mode: bool,
+
+    /// When multiplied against required_min_rx, defines the detection threshold
+    /// connectivity status.
+    ///
+    /// RFC 5880 §6.8.6, a packet with a detect mult of 0 MUST be discarded; we
+    /// encode that here via a `NonZero` type.
+    pub detection_multiplier: NonZeroU8,
+}
+
+impl Default for PeerInfo {
+    fn default() -> Self {
+        Self {
+            // Try to pick a sane default to start with. One second seems
+            // prudent.
+            desired_min_tx: Duration::from_secs(1),
+            required_min_rx: Duration::from_secs(1),
+            discriminator: 0,
+            demand_mode: false,
+            detection_multiplier: DEFAULT_DETECT_MULTIPLIER,
+        }
+    }
+}
+
+impl PeerInfo {
+    /// Initialize a peer info object with a random discriminator.
+    pub fn with_random_discriminator(
+        required_min_rx: Duration,
+        detection_multiplier: NonZeroU8,
+    ) -> Self {
+        Self {
+            required_min_rx,
+            detection_multiplier,
+            discriminator: rand::random(),
+            ..Default::default()
         }
     }
 }
